Fixed the create project tool
